# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for payment sandbox simulation.",
        "title": "Payment Sandbox API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/refunds": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin lists refunds with optional status filter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "refund"
                ],
                "summary": "List refunds",
                "parameters": [
                    {
                        "enum": [
                            "REQUESTED",
                            "APPROVED",
                            "REJECTED",
                            "SUCCESS",
                            "FAILED"
                        ],
                        "type": "string",
                        "description": "Refund status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/payment-sandbox_app_modules_refund_models_entity.Refund"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/refunds/{id}/process": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin processes approved refund to SUCCESS or FAILED",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "refund"
                ],
                "summary": "Process refund",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Refund ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Refund process payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_modules_refund_handlers.ProcessRefundRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/app_modules_refund_handlers.RefundProcessResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/refunds/{id}/review": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin reviews refund request (APPROVE or REJECT)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "refund"
                ],
                "summary": "Review refund",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Refund ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Refund review payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_modules_refund_handlers.ReviewRefundRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/app_modules_refund_handlers.RefundResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app_modules_refund_handlers.ProcessRefundRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "SUCCESS",
                        "FAILED"
                    ],
                    "example": "SUCCESS"
                }
            }
        },
        "app_modules_refund_handlers.RefundProcessResponse": {
            "type": "object",
            "properties": {
                "merchant": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_wallet_models_entity.Merchant"
                },
                "refund": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_refund_models_entity.Refund"
                }
            }
        },
        "app_modules_refund_handlers.RefundResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_number": {
                    "type": "string"
                },
                "merchant_id": {
                    "type": "string"
                },
                "merchant_name": {
                    "type": "string"
                },
                "payment_intent_id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_refund_models_entity.RefundStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "app_modules_refund_handlers.ReviewRefundRequest": {
            "type": "object",
            "required": [
                "decision"
            ],
            "properties": {
                "decision": {
                    "type": "string",
                    "enum": [
                        "APPROVE",
                        "REJECT"
                    ],
                    "example": "APPROVE"
                }
            }
        },
...
        "payment-sandbox_app_modules_refund_models_entity.Refund": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_number": {
                    "type": "string"
                },
                "merchant_id": {
                    "type": "string"
                },
                "merchant_name": {
                    "type": "string"
                },
                "payment_intent_id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_refund_models_entity.RefundStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "payment-sandbox_app_modules_refund_models_entity.RefundStatus": {
            "type": "string",
            "enum": [
                "REQUESTED",
                "APPROVED",
                "REJECTED",
                "SUCCESS",
                "FAILED"
            ],
            "x-enum-varnames": [
                "RefundRequested",
                "RefundApproved",
                "RefundRejected",
                "RefundSuccess",
                "RefundFailed"
            ]
        },
        "payment-sandbox_app_modules_wallet_models_entity.Merchant": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "payment-sandbox_app_shared_response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                },
                "meta": {}
            }
        },
        "payment-sandbox_app_shared_response.ErrorPayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}